make Frame::get_position() retrun type consistent

+++ src/mlt++/MltFrame.cpp
@@ -95,7 +95,7 @@ mlt_properties Frame::get_unique_properties( Service &service )
 	return mlt_frame_unique_properties( get_frame(), service.get_service() );
 }
 
-int64_t Frame::get_position( )
+int Frame::get_position( )
 {
 	return mlt_frame_get_position( get_frame() );
 }
