+++ src/mlt++/MltFrame.h
@@ -47,7 +47,7 @@ namespace Mlt
 			void *get_audio( mlt_audio_format &format, int &frequency, int &channels, int &samples );
 			unsigned char *get_waveform( int w, int h );
 			Producer *get_original_producer( );
-			int64_t get_position( );
+			int get_position( );
 			mlt_properties get_unique_properties( Service &service );
 			int set_image( uint8_t *image, int size, mlt_destructor destroy );
 			int set_alpha( uint8_t *alpha, int size, mlt_destructor destroy );
